Skip generated remote share is exact user match found
Whenever user ids are in the format "user@server", entering these in the share
dialog would also generate a remote share entry. Selecting the remote
entry would usually create a federated share to the same server.

In the case a user enters the full id address as share recipient it
is most likely that the intention was to share with the local user.

This fix removes the automatically generated remote id from the sharees
list when such situation occurs to avoid selecting the wrong entry by
mistake.